requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# code/data/preprocessing_scripts/create_graph.py: 1
networkx == 2.0

# code/data/feed_data.py: 3
# code/data/grapher.py: 3
# code/data/preprocessing_scripts/create_inverse.py: 3
# code/data/preprocessing_scripts/nell.py: 4
# code/data/preprocessing_scripts/preprocess.py: 3
# code/data/preprocessing_scripts/preprocess_countries.py: 3
# code/model/agent.py: 1
# code/model/environment.py: 3
# code/model/trainer.py: 8
numpy == 1.13.3

# code/model/agent.py: 2
# code/model/baseline.py: 3
# code/model/trainer.py: 9
tensorflow_gpu == 1.3.0

# code/data/feed_data.py: 1
# code/model/trainer.py: 3
tqdm == 4.19.4
